Add files via upload
Summation, Factorial and Fibonacci use one number as input,
while GCD, power use two numbers as input. 
And input is a postive number! If an invlaid input is given 
for the functions then the program will tell the user it
is an invalid input, so a negative number will
 give an invalid print statement. 

If any of the inputs are less than 0 then
 in any case it will print out the 
"Invalid input and let the user know"
If the right input is given then the
 order follows as the menu. 
 1 Summation 2 Factorial 3 Fibonacci 4. GCD 
 5. Power and 6 Exits the program!